Refactor : FacultyService

+++ backend/FacultyService/src/main/java/com/example/facultyservice/domain/Faculty.java
@@ -6,11 +6,11 @@ import lombok.Data;
 @Entity
 @Data
 public class Faculty {
-    
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int fId;
-    
+    private Long fId;
+
     private String name;
     private String email;
     private String password;
